Specify the version of http-aws-es needed
`http-aws-es` has had quite a few updates now and it does not support passing through credentials in the same manner. This could be remedied by simply creating a AWS.Config object to pass in as a parameter instead of the credentials however just updating the readme should also make it functional.